// ui/src/generated/AutomationTarget.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DeviceId } from "./DeviceId";
import type { ParamId } from "./ParamId";
import type { SendId } from "./SendId";
import type { TrackId } from "./TrackId";

/**
 * The automated parameter.
 */
export type AutomationTarget = { "type": "TrackVolume", track: TrackId, } | { "type": "TrackPan", track: TrackId, } | { "type": "SendLevel", send: SendId, } | { "type": "DeviceParam", device: DeviceId, param: ParamId, };